Integrating food pages - NavBar index + screen - Moved files to package + data classes - Fix wrong day index
