+++ app/src/main/java/com/zhmu100/ma/ui/components/food/FoodParametersTable.kt
@@ -1,8 +1,16 @@
-package com.zhmu100.ma.ui.components.buttons
+package com.zhmu100.ma.ui.components.food
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
-import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.IntrinsicSize
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -12,6 +20,8 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import com.zhmu100.ma.ui.data.Mineral
+import com.zhmu100.ma.ui.data.Vitamin
 import com.zhmu100.ma.ui.theme.MATheme
 
 /**
@@ -47,11 +57,6 @@ import com.zhmu100.ma.ui.theme.MATheme
  * )
  * ```
  */
-
-
-data class Vitamin(val name: String, val amount: Double, val unit: String)
-data class Mineral(val name: String, val amount: Double, val unit: String)
-
 @Composable
 fun FoodParametersTable(
     calories: Double,
