+++ app/src/main/java/com/zhmu100/ma/ui/components/food/MealEntrySwitcher.kt
@@ -1,15 +1,32 @@
-package com.zhmu100.ma.ui.components.buttons
+package com.zhmu100.ma.ui.components.food
 
 import androidx.annotation.DrawableRes
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Add
 import androidx.compose.material.icons.filled.KeyboardArrowRight
-import androidx.compose.material3.*
-import androidx.compose.runtime.*
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.Divider
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
@@ -20,27 +37,8 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.zhmu100.ma.R
+import com.zhmu100.ma.ui.data.FoodItem
 import com.zhmu100.ma.ui.theme.MATheme
-
-/**
- * Представляет продукт питания, добавленный пользователем в один из приемов пищи.
- *
- * @property name Название продукта.
- * @property amount Количество (в граммах или мл) в текстовом формате.
- * @property fat Количество жиров (в граммах).
- * @property carbs Количество углеводов (в граммах).
- * @property protein Количество белков (в граммах).
- * @property calories Количество калорий.
- */
-
-data class FoodItem(
-    val name: String,
-    val amount: String,
-    val fat: Int,
-    val carbs: Int,
-    val protein: Int,
-    val calories: Int
-)
 
 /**
  * Компонент, отображающий либо кнопку добавления еды, либо карточку с продуктами,
