+++ app/src/main/java/com/zhmu100/ma/ui/data/Mineral.kt
@@ -1,0 +1,3 @@
+package com.zhmu100.ma.ui.data
+
+data class Mineral(val name: String, val amount: Double, val unit: String)
